+++ src/lingdocs/postprocessing.py
@@ -5,7 +5,7 @@ from io import StringIO
 import pandas as pd
 from writio import load
 
-from lingdocs.helpers import load_table_metadata
+from lingdocs.io import load_table_metadata
 from lingdocs.preprocessing import process_metadata
 
 log = logging.getLogger(__name__)
@@ -58,7 +58,7 @@ def insert_tables(md, builder, tables):
     yield md[current:]
 
 
-def postprocess(md_str, builder, dataset, source_dir="."):
+def postprocess(md_str, dataset, builder, source_dir="."):
     tables = process_metadata(load_table_metadata(source_dir), dataset, builder)
     md_str = "".join(insert_manex(md_str, builder, MANPEX_PATTERN, kind="multipart"))
     md_str = "".join(